chore: renaming and restructuring files
