+++ webui/src/Controls/InternalInstanceFields.jsx
@@ -1,6 +1,6 @@
 import React, { useContext, useMemo } from 'react'
-import { DropdownInputField } from '../../Components'
-import { MAX_BUTTONS } from '../../Constants'
+import { DropdownInputField } from '../Components'
+import { MAX_BUTTONS } from '../Constants'
 import {
 	CustomVariableDefinitionsContext,
 	InstancesContext,
@@ -8,7 +8,7 @@ import {
 	SurfacesContext,
 	TriggersContext,
 	VariableDefinitionsContext,
-} from '../../util'
+} from '../util'
 
 export function InternalInstanceField(option, isOnBank, readonly, value, setValue) {
 	switch (option.type) {
