+++ webui/src/Controls/OptionsInputField.jsx
@@ -7,10 +7,10 @@ import {
 	NumberInputField,
 	TextInputField,
 	TextWithVariablesInputField,
-} from '../../Components'
+} from '../Components'
 import { InternalCustomVariableDropdown, InternalInstanceField } from './InternalInstanceFields'
 
-export function ActionTableRowOption({
+export function OptionsInputField({
 	instanceId,
 	isOnBank,
 	isAction,
@@ -113,7 +113,6 @@ export function ActionTableRowOption({
 		}
 		case 'custom-variable': {
 			if (isAction) {
-				//isAction
 				control = (
 					<InternalCustomVariableDropdown disabled={readonly} value={value} setValue={setValue2} includeNone={true} />
 				)
